pyproject.toml
[project]
name = "ancient-coin-detection-system"
version = "0.1.0"
description = "YOLO-based ancient coin detection and knowledge GUI system."
requires-python = ">=3.10,<3.11"
dependencies = [
    "matplotlib==3.10.8",
    "numpy==2.2.6",
    "opencv-python==4.13.0.92",
    "pandas==2.3.3",
    "psutil==7.2.2",
    "pyside6==6.11.0",
    "pyyaml==6.0.3",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "ultralytics==8.4.31",
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
